changing file structure, moving util module to lib/
