+++ lib/email_templates/email_chalaccepted.html
@@ -6,17 +6,17 @@
    <title>Rise To Challenge</title>
 </head>
 
-<body style="width: 100%; max-width: 720px; margin: auto;">
+<body style="width: 100%; max-width: 720px;">
    <main>
-      <h3 style="font-weight: 200;">Hi %username%,</h3>
+      <h3 style="font-weight: 200;">Hi %challenger%,</h3>
       <br>
       <img src="%image%" />
       <br>
-      <p>Unfortunately your proof for the challenge <strong>"%challengename%"</strong> has been rejected by %challenger%, better luck next time. Look to try another challenge? Click on the link below.</p>
+      <p>Your challenge <strong>"%challengename%"</strong> to <strong>%username%</strong> has been accepted. To see the challenge on your dashboard click on the link below.</p>
       <a href="%link%">%linktext%</a>
       <br>
       <br>
-    	<p>Good Luck!!</p>
+    	<p>Enjoy!!</p>
    </main>
    <footer>
       © Rise to Challenge Inc. 320 E 9th St, Charlotte, NC 28202 USA
