+++ lib/email_templates/email_chalrecipient.html
@@ -6,7 +6,7 @@
    <title>Rise To Challenge</title>
 </head>
 
-<body style="width: 100%; max-width: 720px; margin: auto;">
+<body style="width: 100%; max-width: 720px;">
    <main>
       <h3 style="font-weight: 200;">Hi %username%,</h3>
       <br>
