+++ lib/email_templates/email_chalrejected.html
@@ -6,17 +6,17 @@
    <title>Rise To Challenge</title>
 </head>
 
-<body style="width: 100%; max-width: 720px; margin: auto;">
+<body style="width: 100%; max-width: 720px;">
    <main>
       <h3 style="font-weight: 200;">Hi %challenger%,</h3>
       <br>
       <img src="%image%" />
       <br>
-      <p>Your challenge <strong>"%challengename%"</strong> to <strong>%username%</strong> has been accepted. To see the challenge on your dashboard click on the link below.</p>
+      <p>We are sorry to let you know that your challenge <strong>"%challengename%"</strong> has been rejected by %username%, our research suggested that a great video increases the odds of success. To issue more challenge click on the link below.</p>
       <a href="%link%">%linktext%</a>
       <br>
       <br>
-    	<p>Enjoy!!</p>
+    	<p>Good Luck!!</p>
    </main>
    <footer>
       © Rise to Challenge Inc. 320 E 9th St, Charlotte, NC 28202 USA
