+++ lib/email_templates/email_proofrejected.html
@@ -6,13 +6,13 @@
    <title>Rise To Challenge</title>
 </head>
 
-<body style="width: 100%; max-width: 720px; margin: auto;">
+<body style="width: 100%; max-width: 720px;">
    <main>
-      <h3 style="font-weight: 200;">Hi %challenger%,</h3>
+      <h3 style="font-weight: 200;">Hi %username%,</h3>
       <br>
       <img src="%image%" />
       <br>
-      <p>We are sorry to let you know that your challenge <strong>"%challengename%"</strong> has been rejected by %username%, our research suggested that a great video increases the odds of success. To issue more challenge click on the link below.</p>
+      <p>Unfortunately your proof for the challenge <strong>"%challengename%"</strong> has been rejected by %challenger%, better luck next time. Look to try another challenge? Click on the link below.</p>
       <a href="%link%">%linktext%</a>
       <br>
       <br>
