+++ views/sendChallenge.handlebars
@@ -1,8 +1,8 @@
 {{> challenge_nav }}
 <main>
-   <div>
+<!--    <div>
       <h1>Rise-to-the-Challenge</h1>
-   </div>
+   </div> -->
    <div class="z-depth-2 cont">
       <h5>Issue your Challenge</h5>
       <div class="collection Form">
